+++ clang-tools-extra/clang-tidy/readability/ImplicitBoolConversionCheck.cpp
@@ -40,18 +40,18 @@ AST_MATCHER(Stmt, isNULLMacroExpansion) {
 
 StringRef getZeroLiteralToCompareWithForType(CastKind CastExprKind,
                                              QualType Type, ASTContext &Context,
-                                             bool UseUpperCaseSuffix) {
+                                             bool UseUpperCaseLiteralSuffix) {
   switch (CastExprKind) {
   case CK_IntegralToBoolean: {
     if (Type->isUnsignedIntegerType())
-      return UseUpperCaseSuffix ? "0U" : "0u";
+      return UseUpperCaseLiteralSuffix ? "0U" : "0u";
 
     return "0";
   }
 
   case CK_FloatingToBoolean: {
     if (Context.hasSameType(Type, Context.FloatTy)) {
-      return UseUpperCaseSuffix ? "0.0F" : "0.0f";
+      return UseUpperCaseLiteralSuffix ? "0.0F" : "0.0f";
     }
     return "0.0";
   }
@@ -74,7 +74,7 @@ bool isUnaryLogicalNotOperator(const Stmt *Statement) {
 
 void fixGenericExprCastToBool(DiagnosticBuilder &Diag,
                               const ImplicitCastExpr *Cast, const Stmt *Parent,
-                              ASTContext &Context, bool UseUpperCaseSuffix) {
+                              ASTContext &Context, bool UseUpperCaseLiteralSuffix) {
   // In case of expressions like (! integer), we should remove the redundant not
   // operator and use inverted comparison (integer == 0).
   bool InvertComparison =
@@ -121,7 +121,7 @@ void fixGenericExprCastToBool(DiagnosticBuilder &Diag,
   }
 
   EndLocInsertion += getZeroLiteralToCompareWithForType(
-      Cast->getCastKind(), SubExpr->getType(), Context, UseUpperCaseSuffix);
+      Cast->getCastKind(), SubExpr->getType(), Context, UseUpperCaseLiteralSuffix);
 
   if (NeedOuterParens) {
     EndLocInsertion += ")";
@@ -201,7 +201,7 @@ void fixGenericExprCastFromBool(DiagnosticBuilder &Diag,
 
 StringRef getEquivalentForBoolLiteral(const CXXBoolLiteralExpr *BoolLiteral,
                                       QualType DestType, ASTContext &Context,
-                                      bool UseUpperCaseSuffix) {
+                                      bool UseUpperCaseLiteralSuffix) {
   // Prior to C++11, false literal could be implicitly converted to pointer.
   if (!Context.getLangOpts().CPlusPlus11 &&
       (DestType->isPointerType() || DestType->isMemberPointerType()) &&
@@ -212,18 +212,18 @@ StringRef getEquivalentForBoolLiteral(const CXXBoolLiteralExpr *BoolLiteral,
   if (DestType->isFloatingType()) {
     if (Context.hasSameType(DestType, Context.FloatTy)) {
       if (BoolLiteral->getValue())
-        return UseUpperCaseSuffix ? "1.0F" : "1.0f";
+        return UseUpperCaseLiteralSuffix ? "1.0F" : "1.0f";
 
-      return UseUpperCaseSuffix ? "0.0F" : "0.0f";
+      return UseUpperCaseLiteralSuffix ? "0.0F" : "0.0f";
     }
     return BoolLiteral->getValue() ? "1.0" : "0.0";
   }
 
   if (DestType->isUnsignedIntegerType()) {
     if (BoolLiteral->getValue())
-      return UseUpperCaseSuffix ? "1U" : "1u";
+      return UseUpperCaseLiteralSuffix ? "1U" : "1u";
 
-    return UseUpperCaseSuffix ? "0U" : "0u";
+    return UseUpperCaseLiteralSuffix ? "0U" : "0u";
   }
   return BoolLiteral->getValue() ? "1" : "0";
 }
@@ -264,13 +264,13 @@ ImplicitBoolConversionCheck::ImplicitBoolConversionCheck(
     : ClangTidyCheck(Name, Context),
       AllowIntegerConditions(Options.get("AllowIntegerConditions", false)),
       AllowPointerConditions(Options.get("AllowPointerConditions", false)),
-      UseUpperCaseSuffix(Options.get("UseUpperCaseSuffix", false)) {}
+      UseUpperCaseLiteralSuffix(Options.get("UseUpperCaseLiteralSuffix", false)) {}
 
 void ImplicitBoolConversionCheck::storeOptions(
     ClangTidyOptions::OptionMap &Opts) {
   Options.store(Opts, "AllowIntegerConditions", AllowIntegerConditions);
   Options.store(Opts, "AllowPointerConditions", AllowPointerConditions);
-  Options.store(Opts, "UseUpperCaseSuffix", UseUpperCaseSuffix);
+  Options.store(Opts, "UseUpperCaseLiteralSuffix", UseUpperCaseLiteralSuffix);
 }
 
 void ImplicitBoolConversionCheck::registerMatchers(MatchFinder *Finder) {
@@ -395,7 +395,7 @@ void ImplicitBoolConversionCheck::handleCastToBool(const ImplicitCastExpr *Cast,
   if (!EquivalentLiteral.empty()) {
     Diag << tooling::fixit::createReplacement(*Cast, EquivalentLiteral);
   } else {
-    fixGenericExprCastToBool(Diag, Cast, Parent, Context, UseUpperCaseSuffix);
+    fixGenericExprCastToBool(Diag, Cast, Parent, Context, UseUpperCaseLiteralSuffix);
   }
 }
 
@@ -411,7 +411,7 @@ void ImplicitBoolConversionCheck::handleCastFromBool(
           dyn_cast<CXXBoolLiteralExpr>(Cast->getSubExpr()->IgnoreParens())) {
     Diag << tooling::fixit::createReplacement(
         *Cast, getEquivalentForBoolLiteral(BoolLiteral, DestType, Context,
-                                           UseUpperCaseSuffix));
+                                           UseUpperCaseLiteralSuffix));
   } else {
     fixGenericExprCastFromBool(Diag, Cast, Context, DestType.getAsString());
   }