+++ clang-tools-extra/clang-tidy/readability/ImplicitBoolConversionCheck.h
@@ -36,7 +36,7 @@ private:
 
   const bool AllowIntegerConditions;
   const bool AllowPointerConditions;
-  const bool UseUpperCaseSuffix;
+  const bool UseUpperCaseLiteralSuffix;
 };
 
 } // namespace clang::tidy::readability