+++ clang-tools-extra/test/clang-tidy/checkers/readability/implicit-bool-conversion.c
@@ -1,7 +1,7 @@
 // RUN: %check_clang_tidy %s readability-implicit-bool-conversion %t -- -- -std=c23
 // RUN: %check_clang_tidy -check-suffix=UPPER-CASE %s readability-implicit-bool-conversion %t -- \
 // RUN:     -config='{CheckOptions: { \
-// RUN:         readability-implicit-bool-conversion.UseUpperCaseSuffix: true \
+// RUN:         readability-implicit-bool-conversion.UseUpperCaseLiteralSuffix: true \
 // RUN:     }}' -- -std=c23
 
 #undef NULL